MTN: Refactoring helper functions
